Check process id when reloading

+++ module.js
@@ -1,6 +1,6 @@
 var fs = require("fs");
 var cur_pid = parseInt(fs.readFileSync(".muon").toString());
-if (!isNaN(cur_pid)){
+if (cur_pid != process.pid && !isNaN(cur_pid)){
     console.log("Server already running: process id "+cur_pid);
     process.kill();
 }
